+++ src/main/java/pl/kapusta/sdanalysis/stocksource/StockDataProvider.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Map;
 
 public interface StockDataProvider {
-    List<StockData> getData(String stockName, Integer numberOfDays);
+    List<StockData> getData(String stockName, Double numberOfDays);
     Map<String, List<StockData>> getMany(List<String> stockNames, Integer numberOfDays);
     List<CompaniesData> getStockNames();
 }
